Update and rename classes.html to brain_teaser1.html
I am the owner of a school named Chrysland International College, where we have 20 number of students. 5 male and 15 female students. The males and females are arranged in context of both ordered and unordered form. Where the males and females study three different subjects, which are math, physics, and chemistry. The average score for the male children is 70% in maths, while in chemistry 75% average for male, and for physics 86%. While the average for the female is 70% for maths, and 72% for physics, and 90% for chemistry. Deduce the last context, not using ordered or unordered list. but using normal HTML principles