Create brain_teaser6.html
A real estate contractor whose focus is on Lekki environment has over 120 buildings up for sale. Each of these buildings are priced over 15 million naira. 20 of the buildings has colors ranging from pink to brown in a styled manner. The title of the first 5 buildings is Kamil's construction, while the title of the next five is Bigg's construction. Others are Tomi construction and Pat construction respectively. Deduce the brain teaser above