Create brain_teaser8a.html
A software developer  has built over 10 most used web apps on the internet decides to push his files on a reputable web hosting platform. The developer makes use of three different css files. The first one containing the font-families, the second contains the background. The third referencing the body and paragraph. The developer uploads all the necessary files to the server. And renders high quality web based files. Inside the developer's server, consists of an HTML5 file that starts with a DOCTYPE and uses ten different HTML5 tags by default. As the developer kindly this file perspective